fixing task 1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -5,7 +5,7 @@
 * @n: unsigned long int to convert
 */
 
-void print_binary(unsigned int n)
+void print_binary(unsigned long int n)
 {
 if (n > 1)
 print_binary(n >> 1);
